Remove ReviewersAgent from base
