+++ agents/__init__.py
@@ -1,0 +1,12 @@
+from .base import BaseAgent, extract_json_between_markers, get_score
+from .professor import ProfessorAgent
+from .postdoc import PostdocAgent
+from .phd_student import PhDStudentAgent
+from .ml_engineer import MLEngineerAgent
+from .sw_engineer import SWEngineerAgent
+from .reviewers import ReviewersAgent
+__all__ = [
+    'BaseAgent', 'extract_json_between_markers', 'get_score',
+    'ProfessorAgent', 'PostdocAgent', 'PhDStudentAgent',
+    'MLEngineerAgent', 'SWEngineerAgent', 'ReviewersAgent'
+]
