+++ agents/phd_student/__init__.py
@@ -1,0 +1,1 @@
+from .agent import PhDStudentAgent
