+++ agents/postdoc/agent.py
@@ -1,0 +1,79 @@
+from ..base import BaseAgent
+
+
+class PostdocAgent(BaseAgent):
+    def __init__(self, model="gpt4omini", notes=None, max_steps=100, openai_api_key=None):
+        super().__init__(model, notes, max_steps, openai_api_key)
+        self.phases = ["plan formulation", "results interpretation"]
+
+    def context(self, phase):
+        sr_str = str()
+        if self.second_round:
+            sr_str = (
+                f"The following are results from the previous experiments\n",
+                f"Previous Experiment code: {self.prev_results_code}\n"
+                f"Previous Results: {self.prev_exp_results}\n"
+                f"Previous Interpretation of results: {self.prev_interpretation}\n"
+                f"Previous Report: {self.prev_report}\n"
+                f"{self.reviewer_response}\n\n\n"
+            )
+        if phase == "plan formulation":
+            return (
+                sr_str,
+                f"Current Literature Review: {self.lit_review_sum}",
+            )
+        elif phase == "results interpretation":
+            return (
+                sr_str,
+                f"Current Literature Review: {self.lit_review_sum}\n"
+                f"Current Plan: {self.plan}\n"
+                f"Current Dataset code: {self.dataset_code}\n"
+                f"Current Experiment code: {self.results_code}\n"
+                f"Current Results: {self.exp_results}"
+            )
+        return ""
+
+    def example_command(self, phase):
+        if phase not in self.phases:
+            raise Exception(f"Invalid phase: {phase}")
+        return ()
+
+    def command_descriptions(self, phase):
+        if phase not in self.phases:
+            raise Exception(f"Invalid phase: {phase}")
+        if phase == "plan formulation":
+            return (
+                "You can produce dialogue using the following command: ```DIALOGUE\ndialogue here\n```\n where dialogue here is the actual dialogue you will send and DIALOGUE is just the word DIALOGUE.\n"
+                "When you believe a good plan has been arrived at between you and the PhD student you can use the following command to end the dialogue and submit the plan ```PLAN\nplan here\n```\n where plan here is the actual plan to be transmitted and PLAN is just the word PLAN. Plan here should provide a clear outline for how to achieve the task, including what machine learning models to use and implement, what types of datasets should be searched for and used to train the model, and the exact details of the experiment.\n"
+                "You can only use a SINGLE command per inference turn. Do not use more than one command per inference. If you use multiple commands, then only one of them will be executed, NOT BOTH.\n"
+                "Make sure not to produce too much dialogue and to submit an plan in reasonable time."
+                "When performing a command, make sure to include the three ticks (```) at the top and bottom ```COMMAND\ntext\n``` where COMMAND is the specific command you want to run (e.g. PLAN, DIALOGUE).\n"
+            )
+        elif phase == "results interpretation":
+            return (
+                "When you believe a good interpretation has been arrived at between you and the PhD student you can use the following command to end the dialogue and submit the plan ```INTERPRETATION\ninterpretation here\n```\n where interpretation here is the actual interpretation to be transmitted and INTERPRETATION is just the word INTERPRETATION. Please provide an INTERPRETATION in a reasonable amount of time.\n"
+                "You can produce dialogue using the following command: ```DIALOGUE\ndialogue here\n```\n where dialogue here is the actual dialogue you will send and DIALOGUE is just the word DIALOGUE.\n"
+                "You must submit the interpretation during this phase in a reasonable amount of time. Do not delay the submission."
+                "When performing a command, make sure to include the three ticks (```) at the top and bottom ```COMMAND\ntext\n``` where COMMAND is the specific command you want to run (e.g. INTERPRETATION, DIALOGUE).\n"
+            )
+
+    def phase_prompt(self, phase):
+        if phase not in self.phases:
+            raise Exception(f"Invalid phase: {phase}")
+        if phase == "plan formulation":
+            phase_str = (
+                "You are directing a PhD student to help them come up with a good plan, and you interact with them through dialogue.\n"
+                "Your goal is to produce plans that would make good experiments for the given topic. You should aim for a very simple experiment that showcases your plan, not a complex one. You should integrate the provided literature review and come up with plans on how to expand and build on these works for the given topic. Your plans should provide a clear outline for how to achieve the task, including what machine learning models to use and implement, what types of datasets should be searched for and used to train the model, and the exact details of the experiment. Your idea should be very innovative and unlike anything seen before.\n"
+            )
+        elif phase == "results interpretation":
+            phase_str = (
+                "You are directing a PhD student to help them come up with an interpretation for results from an experiment, and you interact with them through dialogue.\n"
+                "Your goal is to interpret results from experiments that were previously run. You should read through the code and look at the results to understand what occurred. You should then discuss with the PhD student how they can interpret the results and give their feedback to improve their thoughts. You should integrate the provided literature review, code, and plans to come up with an exciting interpretation that could make a compelling paper. Your plans should provide a clear outline that can be used to write an academic paper.\n"
+                "Your interpretation should include numbers, relevant metrics to the experiment (e.g. accuracy or loss) and measures of significance. You must propagate this information accurately. You must also complete this in a reasonable amount of time and then submit your results.\n"
+            )
+        return phase_str
+
+    def role_description(self):
+        return "a computer science postdoctoral student at a top university."
+
+
