+++ agents/sw_engineer/agent.py
@@ -1,0 +1,59 @@
+from ..base import BaseAgent
+
+
+class SWEngineerAgent(BaseAgent):
+    def __init__(self, model="gpt4omini", notes=None, max_steps=100, openai_api_key=None):
+        super().__init__(model, notes, max_steps, openai_api_key)
+        self.phases = [
+            "data preparation",
+        ]
+
+    def context(self, phase):
+        sr_str = str()
+        if self.second_round:
+            sr_str = (
+                f"The following are results from the previous experiments\n",
+                f"Previous Experiment code: {self.prev_results_code}\n"
+                f"Previous Results: {self.prev_exp_results}\n"
+                f"Previous Interpretation of results: {self.prev_interpretation}\n"
+                f"Previous Report: {self.prev_report}\n"
+                f"{self.reviewer_response}\n\n\n"
+            )
+        if phase == "data preparation":
+            return (
+                sr_str,
+                f"Current Literature Review: {self.lit_review_sum}\nPlan: {self.plan}",
+                f"Current Plan: {self.plan}")
+        return ""
+
+    def example_command(self, phase):
+        if phase not in self.phases:
+            raise Exception(f"Invalid phase: {phase}")
+        return ()
+
+    def command_descriptions(self, phase):
+        if phase not in self.phases:
+            raise Exception(f"Invalid phase: {phase}")
+        if phase == "data preparation":
+            return (
+                "You can produce dialogue using the following command: ```DIALOGUE\ndialogue here\n```\n where 'dialogue here' is the actual dialogue you will send and DIALOGUE is just the word DIALOGUE.\n"
+                "When you and the ML engineer have finalized your dataset preparation code and are ready to submit the final code, please use the following command: ```SUBMIT_CODE\ncode here\n```\n where 'code here' is the finalized code you will send and SUBMIT_CODE is just the word SUBMIT_CODE. Do not use any classes or functions. The submitted code must have a HuggingFace dataset import and must use an external HuggingFace dataset. If your code returns any errors, they will be provided to you, and you are also able to see print statements.  Make sure function variables are created inside the function or passed as a function parameter. DO NOT CREATE A MAIN FUNCTION.\n"
+                "Make sure to submit code in a reasonable amount of time. Do not make the code too complex, try to make it simple. Do not take too long to submit code. Submit the code early. You should submit the code ASAP.\n"
+                "You can only use a single command per inference turn. Do not use more than one command per inference. If you use multiple commands, then only one of them will be executed, not both.\n"
+                "When performing a command, make sure to include the three ticks (```) at the top and bottom ```COMMAND\ntext\n``` where COMMAND is the specific command you want to run (e.g. SUBMIT_CODE, DIALOGUE).\n")
+        return ""
+
+    def phase_prompt(self, phase):
+        if phase not in self.phases:
+            raise Exception(f"Invalid phase: {phase}")
+        elif phase == "data preparation":
+            phase_str = (
+                "You are a software engineer directing a machine learning engineer, where the machine learning engineer will be writing the code, and you can interact with them through dialogue.\n"
+                "Your goal is to help the ML engineer produce code that prepares the data for the provided experiment. You should aim for very simple code to prepare the data, not complex code. You should integrate the provided literature review and the plan and come up with code to prepare data for this experiment.\n"
+            )
+        return phase_str
+
+    def role_description(self):
+        return "a software engineer working at a top university."
+
+
